requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
supabase>=2.7,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
httpx>=0.27,<1.0
requests>=2.32,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
joblib>=1.4,<2.0
apscheduler>=3.10,<4.0
xgboost>=2.1,<3.0
